Refactor action slice helpers into dedicated modules
