+++ src/core/state/slices/actions/index.js
@@ -1,0 +1,2 @@
+export { getInstanceProgressSnapshot } from './progress.js';
+export { ensureSlice, getSliceState } from './registry.js';
