+++ src/core/state/slices/actions/instances.js
@@ -1,13 +1,17 @@
-import { structuredClone, createId, toNumber } from '../../helpers.js';
-import { getActionDefinition, getHustleDefinition } from '../registry.js';
-import knowledgeTrackData from '../../../game/requirements/data/knowledgeTracks.js';
-import { createRegistrySliceManager } from './factory.js';
+import { structuredClone, createId, toNumber } from '../../../helpers.js';
+import knowledgeTrackData from '../../../../game/requirements/data/knowledgeTracks.js';
+import { getActionDefinition, getHustleDefinition } from '../../registry.js';
+import { normalizeInstanceProgress } from './progress.js';
 
 const KNOWLEDGE_TRACKS = knowledgeTrackData;
-const MAX_ACTION_INSTANCES = 100;
-const COMPLETED_RETENTION_DAYS = 1;
+export const MAX_ACTION_INSTANCES = 100;
+export const COMPLETED_RETENTION_DAYS = 1;
 
-function isInstanceCompleted(instance) {
+export function resolveDefinition(id) {
+  return getActionDefinition(id) || getHustleDefinition(id);
+}
+
+export function isInstanceCompleted(instance) {
   if (!instance || typeof instance !== 'object') {
     return false;
   }
@@ -22,7 +26,7 @@ function isInstanceCompleted(instance) {
   return false;
 }
 
-function resolveCompletionDay(instance) {
+export function resolveCompletionDay(instance) {
   if (!instance || typeof instance !== 'object') {
     return null;
   }
@@ -41,7 +45,7 @@ function resolveCompletionDay(instance) {
   return null;
 }
 
-function shouldRetireInstance(instance, currentDay) {
+export function shouldRetireInstance(instance, currentDay) {
   const completionDay = resolveCompletionDay(instance);
   if (!Number.isFinite(currentDay) || completionDay == null) {
     return false;
@@ -49,247 +53,7 @@ function shouldRetireInstance(instance, currentDay) {
   return currentDay - completionDay >= COMPLETED_RETENTION_DAYS;
 }
 
-function resolveDefinition(id) {
-  return getActionDefinition(id) || getHustleDefinition(id);
-}
-
-function normalizeProgressLog(log = {}) {
-  const normalized = {};
-  for (const [dayKey, value] of Object.entries(log || {})) {
-    const day = Math.max(1, Math.floor(Number(dayKey)) || 1);
-    const hours = Number(value);
-    if (!Number.isFinite(hours) || hours < 0) continue;
-    normalized[day] = Number.isFinite(normalized[day]) ? normalized[day] + hours : hours;
-  }
-  return normalized;
-}
-
-function normalizeInstanceProgress(definition, instance = {}) {
-  const template = typeof definition?.progress === 'object' && definition.progress !== null
-    ? definition.progress
-    : {};
-  const source = typeof instance.progress === 'object' && instance.progress !== null
-    ? instance.progress
-    : {};
-
-  const progress = {};
-  progress.type = typeof source.type === 'string' ? source.type : template.type || 'instant';
-  const fallbackCompletion = template.completion || (progress.type === 'instant' ? 'instant' : 'deferred');
-  progress.completion = typeof source.completion === 'string' ? source.completion : fallbackCompletion;
-
-  const hoursRequiredSource = toNumber(source.hoursRequired, null);
-  const hoursRequiredTemplate = toNumber(template.hoursRequired, null);
-  const hoursRequiredInstance = toNumber(instance.hoursRequired, null);
-  const resolvedHoursRequired = hoursRequiredSource != null
-    ? hoursRequiredSource
-    : hoursRequiredTemplate != null
-      ? hoursRequiredTemplate
-      : hoursRequiredInstance;
-  progress.hoursRequired = Number.isFinite(resolvedHoursRequired) && resolvedHoursRequired >= 0
-    ? resolvedHoursRequired
-    : null;
-
-  const hoursPerDaySource = toNumber(source.hoursPerDay, null);
-  const hoursPerDayTemplate = toNumber(template.hoursPerDay, null);
-  const resolvedHoursPerDay = hoursPerDaySource != null ? hoursPerDaySource : hoursPerDayTemplate;
-  progress.hoursPerDay = Number.isFinite(resolvedHoursPerDay) && resolvedHoursPerDay > 0 ? resolvedHoursPerDay : null;
-
-  const daysRequiredSource = toNumber(source.daysRequired, null);
-  const daysRequiredTemplate = toNumber(template.daysRequired, null);
-  const resolvedDaysRequired = daysRequiredSource != null ? daysRequiredSource : daysRequiredTemplate;
-  progress.daysRequired = Number.isFinite(resolvedDaysRequired) && resolvedDaysRequired > 0
-    ? Math.floor(resolvedDaysRequired)
-    : null;
-
-  let deadline = source.deadlineDay;
-  if (deadline == null) {
-    deadline = template.deadlineDay;
-  }
-  if (deadline == null) {
-    deadline = instance.deadlineDay;
-  }
-  progress.deadlineDay = Number.isFinite(deadline) && deadline > 0 ? Math.floor(deadline) : null;
-
-  const normalizedLog = normalizeProgressLog(source.dailyLog);
-  progress.dailyLog = normalizedLog;
-  let totalHours = 0;
-  let lastWorked = null;
-  for (const [dayKey, value] of Object.entries(normalizedLog)) {
-    const day = Math.max(1, Math.floor(Number(dayKey)) || 1);
-    const hours = Number(value);
-    if (!Number.isFinite(hours) || hours < 0) continue;
-    totalHours += hours;
-    if (lastWorked == null || day > lastWorked) {
-      lastWorked = day;
-    }
-  }
-
-  const hoursLoggedSource = toNumber(source.hoursLogged, null);
-  const hoursLoggedInstance = toNumber(instance.hoursLogged, null);
-  const resolvedHoursLogged = hoursLoggedSource != null
-    ? hoursLoggedSource
-    : Number.isFinite(totalHours)
-      ? totalHours
-      : hoursLoggedInstance;
-  progress.hoursLogged = Number.isFinite(resolvedHoursLogged) && resolvedHoursLogged >= 0 ? resolvedHoursLogged : 0;
-
-  const hoursPerDay = Number(progress.hoursPerDay);
-  if (Number.isFinite(hoursPerDay) && hoursPerDay > 0) {
-    const threshold = hoursPerDay - 0.0001;
-    progress.daysCompleted = Object.values(normalizedLog).filter(hours => Number(hours) >= threshold).length;
-  } else {
-    const daysCompletedSource = toNumber(source.daysCompleted, null);
-    progress.daysCompleted = Number.isFinite(daysCompletedSource) && daysCompletedSource >= 0
-      ? Math.floor(daysCompletedSource)
-      : 0;
-  }
-
-  if (lastWorked != null) {
-    progress.lastWorkedDay = lastWorked;
-  } else if (source.lastWorkedDay != null) {
-    const parsed = Math.floor(Number(source.lastWorkedDay));
-    progress.lastWorkedDay = Number.isFinite(parsed) && parsed > 0 ? parsed : null;
-  } else {
-    progress.lastWorkedDay = null;
-  }
-
-  progress.completed = source.completed === true || instance.completed === true;
-  if (source.completedOnDay != null) {
-    const parsed = Math.floor(Number(source.completedOnDay));
-    progress.completedOnDay = Number.isFinite(parsed) && parsed > 0 ? parsed : undefined;
-  } else if (progress.completed && instance.completedOnDay != null) {
-    const parsed = Math.floor(Number(instance.completedOnDay));
-    progress.completedOnDay = Number.isFinite(parsed) && parsed > 0 ? parsed : undefined;
-  }
-
-  return progress;
-}
-
-export function getInstanceProgressSnapshot(instance = {}) {
-  if (!instance || typeof instance !== 'object') {
-    return null;
-  }
-
-  const progress = typeof instance.progress === 'object' && instance.progress !== null
-    ? instance.progress
-    : {};
-
-  const snapshot = {};
-
-  snapshot.definitionId = typeof instance.definitionId === 'string'
-    ? instance.definitionId
-    : typeof progress.definitionId === 'string'
-      ? progress.definitionId
-      : null;
-
-  snapshot.instanceId = typeof instance.id === 'string'
-    ? instance.id
-    : typeof progress.instanceId === 'string'
-      ? progress.instanceId
-      : null;
-
-  const resolvedHoursLogged = toNumber(progress.hoursLogged, toNumber(instance.hoursLogged, 0));
-  snapshot.hoursLogged = Number.isFinite(resolvedHoursLogged) && resolvedHoursLogged >= 0
-    ? resolvedHoursLogged
-    : 0;
-
-  const hoursRequiredCandidates = [
-    progress.hoursRequired,
-    instance.hoursRequired,
-    progress.totalHours
-  ];
-  let hoursRequired = null;
-  for (const candidate of hoursRequiredCandidates) {
-    const numeric = toNumber(candidate, null);
-    if (Number.isFinite(numeric) && numeric >= 0) {
-      hoursRequired = numeric;
-      break;
-    }
-  }
-  snapshot.hoursRequired = hoursRequired;
-
-  snapshot.hoursRemaining = snapshot.hoursRequired != null
-    ? Math.max(0, snapshot.hoursRequired - snapshot.hoursLogged)
-    : null;
-
-  const parsedHoursPerDay = toNumber(progress.hoursPerDay, null);
-  snapshot.hoursPerDay = Number.isFinite(parsedHoursPerDay) && parsedHoursPerDay > 0
-    ? parsedHoursPerDay
-    : null;
-
-  const parsedDaysCompleted = toNumber(progress.daysCompleted, null);
-  snapshot.daysCompleted = Number.isFinite(parsedDaysCompleted) && parsedDaysCompleted >= 0
-    ? Math.floor(parsedDaysCompleted)
-    : 0;
-
-  const parsedDaysRequired = toNumber(progress.daysRequired, null);
-  snapshot.daysRequired = Number.isFinite(parsedDaysRequired) && parsedDaysRequired > 0
-    ? Math.floor(parsedDaysRequired)
-    : null;
-
-  const progressType = typeof progress.type === 'string' && progress.type.trim()
-    ? progress.type.trim()
-    : null;
-  snapshot.type = progressType;
-
-  const completionCandidates = [progress.completion, progress.completionMode];
-  let completionMode = null;
-  for (const candidate of completionCandidates) {
-    if (typeof candidate === 'string' && candidate.trim()) {
-      completionMode = candidate.trim();
-      break;
-    }
-  }
-  if (!completionMode) {
-    completionMode = progressType === 'instant' ? 'instant' : 'manual';
-  }
-  snapshot.completion = completionMode;
-  snapshot.completionMode = completionMode;
-
-  snapshot.percentComplete = snapshot.hoursRequired && snapshot.hoursRequired > 0
-    ? Math.max(0, Math.min(1, snapshot.hoursLogged / snapshot.hoursRequired))
-    : null;
-
-  const parsedLastWorked = toNumber(progress.lastWorkedDay, null);
-  snapshot.lastWorkedDay = Number.isFinite(parsedLastWorked) && parsedLastWorked > 0
-    ? Math.floor(parsedLastWorked)
-    : null;
-
-  const parsedDeadline = toNumber(progress.deadlineDay, null);
-  snapshot.deadlineDay = Number.isFinite(parsedDeadline) && parsedDeadline > 0
-    ? Math.floor(parsedDeadline)
-    : null;
-
-  const parsedAcceptedOn = toNumber(progress.acceptedOnDay ?? instance.acceptedOnDay, null);
-  snapshot.acceptedOnDay = Number.isFinite(parsedAcceptedOn) && parsedAcceptedOn > 0
-    ? Math.floor(parsedAcceptedOn)
-    : null;
-
-  const completed = progress.completed === true
-    || instance.completed === true
-    || progress.status === 'completed'
-    || instance.status === 'completed';
-  snapshot.completed = completed;
-
-  const parsedCompletedOn = toNumber(progress.completedOnDay ?? instance.completedOnDay, null);
-  snapshot.completedOnDay = Number.isFinite(parsedCompletedOn) && parsedCompletedOn > 0
-    ? Math.floor(parsedCompletedOn)
-    : completed
-      ? snapshot.acceptedOnDay
-      : null;
-
-  snapshot.metadata = typeof progress.metadata === 'object' && progress.metadata !== null
-    ? progress.metadata
-    : {};
-
-  if (progress.dailyLog && typeof progress.dailyLog === 'object') {
-    snapshot.dailyLog = progress.dailyLog;
-  }
-
-  return snapshot;
-}
-
-function createDefaultActionState(definition) {
+export function createDefaultActionState(definition) {
   if (!definition) {
     return { instances: [] };
   }
@@ -302,7 +66,7 @@ function createDefaultActionState(definition) {
   return base;
 }
 
-function normalizeActionInstance(definition, instance = {}, { state } = {}) {
+export function normalizeActionInstance(definition, instance = {}, { state } = {}) {
   const base = typeof instance === 'object' && instance !== null ? { ...instance } : {};
   const fallbackDay = Math.max(1, Math.floor(Number(state?.day) || 1));
   if (!base.id) {
@@ -362,7 +126,7 @@ function normalizeActionInstance(definition, instance = {}, { state } = {}) {
   return base;
 }
 
-function normalizeActionState(definition, entry = {}, context) {
+export function normalizeActionState(definition, entry = {}, context) {
   const defaults = createDefaultActionState(definition);
   const normalized = { ...defaults };
   const existing = typeof entry === 'object' && entry !== null ? entry : {};
@@ -428,7 +192,7 @@ function normalizeActionState(definition, entry = {}, context) {
   return normalized;
 }
 
-function buildLegacyStudyInstance(definition, track, progress, state) {
+export function buildLegacyStudyInstance(definition, track, progress, state) {
   if (!definition || !track || !progress) {
     return null;
   }
@@ -493,7 +257,7 @@ function buildLegacyStudyInstance(definition, track, progress, state) {
   return normalizeActionInstance(definition, overrides, { state });
 }
 
-function seedKnowledgeStudyInstances({ state, sliceState }) {
+export function seedKnowledgeStudyInstances({ state, sliceState }) {
   const knowledge = state?.progress?.knowledge;
   if (!knowledge || typeof knowledge !== 'object') {
     return;
@@ -523,7 +287,7 @@ function seedKnowledgeStudyInstances({ state, sliceState }) {
   }
 }
 
-function migrateLegacyHustleProgress({ state, sliceState }) {
+export function migrateLegacyHustleProgress({ state, sliceState }) {
   const legacy = state?.hustles;
   if (legacy && typeof legacy === 'object') {
     for (const [id, legacyEntry] of Object.entries(legacy)) {
@@ -564,14 +328,3 @@ function migrateLegacyHustleProgress({ state, sliceState }) {
 
   seedKnowledgeStudyInstances({ state, sliceState });
 }
-
-const { ensureSlice, getSliceState } = createRegistrySliceManager({
-  sliceKey: 'actions',
-  registryKey: 'actions',
-  definitionLookup: resolveDefinition,
-  defaultFactory: definition => createDefaultActionState(definition),
-  normalizer: (definition, entry, context) => normalizeActionState(definition, entry, context),
-  ensureHook: migrateLegacyHustleProgress
-});
-
-export { ensureSlice, getSliceState };
